Refactor for clarity and security. Modify golangci options. Add godoc to some tokenauthority methods. Create helper functions to manipulate userid field inside context in type-safe way. Create new handler functions to serve openapi.json and swagger ui. Change requestCreateFood fields to int32 from int to be in line with database & json.

+++ util/typeconvert/pgtype_converters.go
@@ -21,7 +21,7 @@ func PgtypeInt4ToInt(i pgtype.Int4) int32 {
 	if !i.Valid {
 		return 0
 	}
-	return int32(i.Int32)
+	return i.Int32
 }
 
 func Float32ToPgtypeFloat4(f float32) pgtype.Float4 {
